Replaced [PATH_TRANSLATED] with [REQUEST_URI]
It initially didn’t work for me using PATH_TRANSLATED, and I found a
note in $_SERVER documentation:

Note: As of PHP 4.3.2, PATH_TRANSLATED is no longer set implicitly
under the Apache 2.

I tried changing my server settings, and nothing came to fruition.

So here’s an alternative - it asks for you to tell it how to navigate
to the top-level from the script, and then REQUEST_URI takes it from
there.

Now it’s working on my server, and I’m very happy!!! :) It’s exactly
what I wanted. Thank you ubergeek42 for making it! <3